+++ src/main/java/com/binance/client/model/event/AggregateTradeEvent.java
@@ -9,21 +9,21 @@ public class AggregateTradeEvent {
 
     private String eventType;
 
-    private Integer eventTime;
+    private Long eventTime;
 
     private String symbol;
 
-    private Integer id;
+    private Long id;
 
     private BigDecimal price;
 
     private BigDecimal qty;
 
-    private Integer firstId;
+    private Long firstId;
 
-    private Integer lastId;
+    private Long lastId;
 
-    private Integer time;
+    private Long time;
 
     private Boolean isBuyerMaker;
 
@@ -37,11 +37,11 @@ public class AggregateTradeEvent {
         this.eventType = eventType;
     }
 
-    public Integer getEventTime() {
+    public Long getEventTime() {
         return eventTime;
     }
 
-    public void setEventTime(Integer eventTime) {
+    public void setEventTime(Long eventTime) {
         this.eventTime = eventTime;
     }
 
@@ -53,11 +53,11 @@ public class AggregateTradeEvent {
         this.symbol = symbol;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -77,27 +77,27 @@ public class AggregateTradeEvent {
         this.qty = qty;
     }
 
-    public Integer getFirstId() {
+    public Long getFirstId() {
         return firstId;
     }
 
-    public void setFirstId(Integer firstId) {
+    public void setFirstId(Long firstId) {
         this.firstId = firstId;
     }
 
-    public Integer getLastId() {
+    public Long getLastId() {
         return lastId;
     }
 
-    public void setLastId(Integer lastId) {
+    public void setLastId(Long lastId) {
         this.lastId = lastId;
     }
 
-    public Integer getTime() {
+    public Long getTime() {
         return time;
     }
 
-    public void setTime(Integer time) {
+    public void setTime(Long time) {
         this.time = time;
     }
